requirements.txt
geopandas==1.1.3
matplotlib==3.10.0
matplotlib-inline==0.2.1
matplotlib-venn==1.1.2
numpy==2.0.2
pandas==2.2.2
pandas-datareader==0.10.0
pandas-gbq==0.30.0
pandas-stubs==2.2.2.240909
polars==1.35.2
polars-runtime-32==1.35.2
PyYAML==6.0.3
scikit-learn==1.6.1
scipy==1.16.3
sklearn-pandas==2.2.0
statsmodels==0.14.6
yfinance==0.2.66
